+++ src/detectors/builtin/branchDuplicate.ts
@@ -11,7 +11,7 @@ import {
   AstStatement,
 } from "../../internals/tact/imports";
 import { SrcInfo } from "../../internals/tact/imports";
-import { MistiTactWarning, Severity } from "../../internals/warnings";
+import { Category, MistiTactWarning, Severity } from "../../internals/warnings";
 import { AstDetector } from "../detector";
 
 /**
@@ -43,6 +43,7 @@ import { AstDetector } from "../detector";
  */
 export class BranchDuplicate extends AstDetector {
   severity = Severity.HIGH;
+  category = Category.SECURITY;
 
   async check(cu: CompilationUnit): Promise<MistiTactWarning[]> {
     return cu.ast.getProgramEntries().reduce((acc, node) => {