+++ src/detectors/builtin/duplicatedCondition.ts
@@ -12,7 +12,7 @@ import {
   prettyPrint,
   AstExpression,
 } from "../../internals/tact/imports";
-import { MistiTactWarning, Severity } from "../../internals/warnings";
+import { Category, MistiTactWarning, Severity } from "../../internals/warnings";
 import { AstDetector } from "../detector";
 
 /**
@@ -46,6 +46,7 @@ import { AstDetector } from "../detector";
  */
 export class DuplicatedCondition extends AstDetector {
   severity = Severity.HIGH;
+  category = Category.SECURITY;
 
   async check(cu: CompilationUnit): Promise<MistiTactWarning[]> {
     return cu.ast.getProgramEntries().reduce((acc, node) => {