+++ src/detectors/builtin/implicitOpcode.ts
@@ -1,5 +1,5 @@
 import { CompilationUnit } from "../../internals/ir";
-import { MistiTactWarning, Severity } from "../../internals/warnings";
+import { Category, MistiTactWarning, Severity } from "../../internals/warnings";
 import { AstDetector } from "../detector";
 import { idText } from "@tact-lang/compiler";
 
@@ -31,6 +31,7 @@ import { idText } from "@tact-lang/compiler";
  */
 export class ImplicitOpcode extends AstDetector {
   severity = Severity.INFO;
+  category = Category.BEST_PRACTICES;
 
   async check(cu: CompilationUnit): Promise<MistiTactWarning[]> {
     return Array.from(cu.ast.getMessages()).reduce((acc, msg) => {