+++ src/detectors/builtin/preferBinaryReceiver.ts
@@ -1,5 +1,5 @@
 import { CompilationUnit } from "../../internals/ir";
-import { MistiTactWarning, Severity } from "../../internals/warnings";
+import { Category, MistiTactWarning, Severity } from "../../internals/warnings";
 import { AstDetector } from "../detector";
 
 /**
@@ -26,6 +26,7 @@ import { AstDetector } from "../detector";
  */
 export class PreferBinaryReceiver extends AstDetector {
   severity = Severity.LOW;
+  category = Category.OPTIMIZATION;
 
   async check(cu: CompilationUnit): Promise<MistiTactWarning[]> {
     return Array.from(cu.ast.getFunctions()).reduce((acc, node) => {