+++ src/detectors/builtin/suspiciousMessageMode.ts
@@ -6,7 +6,7 @@ import {
   AstOpBinary,
 } from "../../internals/tact/imports";
 import { idText } from "../../internals/tact/imports";
-import { MistiTactWarning, Severity } from "../../internals/warnings";
+import { Category, MistiTactWarning, Severity } from "../../internals/warnings";
 import { AstDetector } from "../detector";
 
 /**
@@ -41,6 +41,7 @@ import { AstDetector } from "../detector";
  */
 export class SuspiciousMessageMode extends AstDetector {
   severity = { min: Severity.LOW, max: Severity.MEDIUM };
+  category = Category.SECURITY;
 
   async check(cu: CompilationUnit): Promise<MistiTactWarning[]> {
     const warnings: MistiTactWarning[] = [];